+++ src/XRT.jl
@@ -350,6 +350,7 @@ end
 
 export size, length, setindex!, getindex, wait
 export sync!, group_id, set_arg!, start, stop, load_xclbin!, get_xclbin_uuid, Xclbin
+export hw_context
 export @prepare_bitstream, @prepare_run
 export ToDeviceArray, FromDeviceArray